+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Upgrade dependencies (#241)
+
 ## [0.45.0] - 2023-09-07
 
 ### Added