+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.45.1] - 2023-10-02
+## [0.45.1] - 2023-10-06
 
 ### Changed